+++ alacritty/src/window_context.rs
@@ -231,6 +231,8 @@ impl WindowContext {
         let window_config = &old_config.window;
         if window_config.padding(1.) != self.config.window.padding(1.)
             || window_config.dynamic_padding != self.config.window.dynamic_padding
+            || window_config.valign != self.config.window.valign
+            || window_config.halign != self.config.window.halign
         {
             self.display.pending_update.dirty = true;
         }